rename geo/radar_grid.py
+ reference_radar_grid --> radar_grid.py

+ geogrid.py --> geo_grid.py
